+++ src/compiler.c
@@ -265,7 +265,7 @@ void compile(ConcoctNodeTree* tree)
         fprintf(stderr, "Unable to handle token: %s\n", cct_token_type_to_string(current->token.type));
         break;
     }
-    for(int i = 0; i < current->child_count; i++)
+    for(size_t i = 0; i < current->child_count; i++)
       enqueue(pqueue, current->children[i]);
   }
